requirements.txt
numpy
cupy
progressbar2
scipy
pytest
cProfile
